requirements.txt
aws-cdk-lib==2.162.0
constructs>=10.0.0,<11.0.0
aws_cdk.aws_lambda_python_alpha==2.162.0a0
cdk-nag==2.32.2
setuptools>=70.0.0
